modify git.py file

+++ git.py
@@ -2,33 +2,37 @@ import os
 
 
 def set_proxy(username, password, host_ip, host_port, pc_username):
-    file_name = '/home/'+ pc_username + '/.gitconfig'
-    file_temp = open(file_name + '_temp', 'w+')
-    file_ob = open(file_name, 'r')
-    for line in file_ob.readlines():
-        if 'http' in line or 'https' in line or 'proxy' in line:
-            continue    
-        else:
-            file_temp.write(line)
+    check = os.path.exists('/etc/' + pc_username + '/.gitconfig')
+    if check == True :
+        file_name = '/home/'+ pc_username + ''
+        file_temp = open(file_name + '_temp', 'w+')
+        file_ob = open(file_name, 'r')
+        for line in file_ob.readlines():
+            if 'http' in line or 'https' in line or 'proxy' in line:
+                continue    
+            else:
+                file_temp.write(line)
 
-    file_temp.write('\n[http]\n\tproxy = http://'+ username + ':'+ password + '@' + host_ip +':' +  host_port + '\n')
-    file_temp.write('[https]\n\tproxy=http://'+ username + ':'+ password + '@' + host_ip + ':' + host_port + '\n' )
-    file_ob.close()
-    file_temp.close()
-    os.remove(file_name)
-    os.rename(file_name +'_temp', file_name) 
+        file_temp.write('\n[http]\n\tproxy = http://'+ username + ':'+ password + '@' + host_ip +':' +  host_port + '\n')
+        file_temp.write('[https]\n\tproxy=http://'+ username + ':'+ password + '@' + host_ip + ':' + host_port + '\n' )
+        file_ob.close()
+        file_temp.close()
+        os.remove(file_name)
+        os.rename(file_name +'_temp', file_name) 
 
 def unset_proxy(pc_username):
-    file_name = '/home/'+ pc_username + '/.gitconfig'
-    file_temp = open(file_name + '_temp', 'w+')
-    file_ob = open(file_name, 'r')
-    for line in file_ob.readlines():
-        if 'http' in line or 'https' in line or 'proxy' in line:
-            continue    
-        else:
-            file_temp.write(line)
+    check = os.path.exists('/etc/' + pc_username + '/.gitconfig')
+    if check == True :
+        file_name = '/home/'+ pc_username + '/.gitconfig'
+        file_temp = open(file_name + '_temp', 'w+')
+        file_ob = open(file_name, 'r')
+        for line in file_ob.readlines():
+            if 'http' in line or 'https' in line or 'proxy' in line:
+                continue    
+            else:
+                file_temp.write(line)
 
-    file_ob.close()
-    file_temp.close()
-    os.remove(file_name)
-    os.rename(file_name + '_temp', file_name) 
+        file_ob.close()
+        file_temp.close()
+        os.remove(file_name)
+        os.rename(file_name + '_temp', file_name)
